fixed null exception

+++ Windows/Setup-LeagueClient.ps1
@@ -71,7 +71,11 @@ function Stop-RiotProcesses {
     Stop-Process -Name 'RiotClientUx' -ErrorAction Ignore
     Stop-Process -Name 'LeagueClient' -ErrorAction Ignore
     Remove-Item $RCS_LOCKFILE -Force -ErrorAction Ignore
-    Remove-Item $LCU_LOCKFILE -Force -ErrorAction Ignore
+    
+    if ($LCU_LOCKFILE -ne $null) {
+        Remove-Item $LCU_LOCKFILE -Force -ErrorAction Ignore
+    }
+
     Start-Sleep 5 # Wait for processes to settle.
 }
